separate common part
